requirements.txt
tensorflow>=2.7.0
numpy==1.22.1
Pillow==9.0.0
pytest==6.2.5
matplotlib==3.5.1
opencv==4.5.5